added minimarker feature to 'view latest image'; consolidated some functions
