+++ cyan_angular/src/app/latest-image/latest-image.component.html
@@ -30,7 +30,12 @@
 
   <!-- Leaflet Map -->
   <div class="mini_map_block">
-		<div class="mini_map" leaflet [leafletOptions]="options" (leafletMapReady)="onMapReady($event)"></div>
+		<div class="mini_map"
+			leaflet
+			[leafletOptions]="options"
+			(leafletMapReady)="onMapReady($event)"
+			(leafletClick)="addMarkerOnClick($event)"
+		>
 	</div>
 
 </div>
